+++ History.md
@@ -1,3 +1,5 @@
+* jchoover: FEAT#4194 - Move DownloadUrl functionality from engine to dutil.
+
 * RobMen: WiX v3.9
 
 ## WixBuild: Version 3.9.0.0